requirements.txt
annotated-types
asgiref
autocommand
certifi
charset-normalizer
comtypes
crispy-bootstrap5
distlib
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-environ
django-filter
djangorestframework
env
environ
filelock
inflect
jaraco.context
jaraco.env
jaraco.functools
jaraco.text
Markdown
more-itertools
MouseInfo
packaging
path
Pillow
pip-run
platformdirs
psycopg2
psycopg2-binary
PyAutoGUI
pydantic
pydantic_core
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
pytz
sqlparse
typing_extensions
tzdata
virtualenv